# refactor(node): improve body streaming