This PR improves the streamBody functions performance and closes the stream after it's completed.

This change is strongly inspired by [@honojs/node-server](https://github.com/honojs/node-server) approach.

Before:
```
Node.js:         v22.13.0
OS:              darwin arm64
OHA:             1.6.0


node... 116485 req/sec
srvx... 62466 req/sec
srvx-fast... 101216 req/sec
┌───────────┬──────────────────┐
│ (index)   │ Values           │
├───────────┼──────────────────┤
│ node      │ '116485 req/sec' │
│ srvx-fast │ '101216 req/sec' │
│ srvx      │ '62466 req/sec'  │
└───────────┴──────────────────┘
```

After:
```
Node.js:         v22.13.0
OS:              darwin arm64
OHA:             1.6.0


node... 115214 req/sec
srvx... 66832 req/sec
srvx-fast... 101813 req/sec
┌───────────┬──────────────────┐
│ (index)   │ Values           │
├───────────┼──────────────────┤
│ node      │ '115214 req/sec' │
│ srvx-fast │ '101813 req/sec' │
│ srvx      │ '66832 req/sec'  │
└───────────┴──────────────────┘
```
<!--
PLEASE DO THIS BEFORE SUBMITTING A PR

1) Make sure there is an issue covering the problem or idea first. If not, please create one. Reference it in the PR via "resolves #12312312" 
2) Please keep your changes minimal and split them if you need to.
3) Ensure there is a minimal reproduction attached for bug fixes.

This will greatly help speed up the review process.

Thanks for your contribution ❤️
-->