@pi0 2025-01-15
Thank you for PR ❤️ I did run bench on few servers, it seems difference is not reliable (specially on `-fast` variant, simple loop is sometimes faster in order of 10,000 (second to the left digit))

However I like new impl because socket is closed and we respect drain event 👍🏼